+++ applications/minotari_ledger_wallet/wallet/src/crypto/mod.rs
