+++ applications/minotari_ledger_wallet/wallet/src/handlers/get_one_sided_metadata_signature.rs
@@ -30,7 +30,7 @@ use zeroize::Zeroizing;
 use crate::{
     alloc::string::ToString,
     hashing::DomainSeparatedConsensusHasher,
-    tari_crypto::{
+    crypto::{
         commitment::PedersenCommitment,
         commitment_and_public_key_signature::CommitmentAndPublicKeySignature,
         commitment_factory::PedersenCommitmentFactory,
